crawler: Remove waitForLowActiveIO
Only use dynamic delays for the crawler. Even though max wait was 1 second the number of waits could severely impact crawler speed.

Instead of relying on a global metric we use the stateless local delays to keep the crawler running at a speed more adjusted to current conditions.

The only case we keep it is before bitrot checks, when enabled.