metacache: Allow prefix filtering
Do listings with prefix filter when bloom filter is dirty.

This will forward the prefix filter to the lister which will make it only scan the folders/objects with the specified prefix.

If we have a clean bloom filter we try to build a more generally useful cache so in that case we will list all objects/folders.