Show Delete replication status header
X-Minio-Replication-Delete-Status header shows the
status of the replication of a permanent delete of a version.
All GETs are disallowed and return 405 on this object version.
In the case of replicating delete markers,
X-Minio-Replication-DeleteMarker-Status shows status of replication,
and would similarly return 405.

Additionally, this PR adds reporting of deletemarker event completion
and updates documentation