heal: Return ObjectNotFound in HealObject if the object is not found
erasureServeSets.HealObject() does not know in which set a given
object exists, so it relies on HealObject ObjectNotFound error to go to
the next set in order to find/heal the object. However this behavior
has been changed recently and we can see sometimes many gray objects
with mc admin heal command.

This commit reverts the original logic.