parentDirIsObject() to return quickly with inexistant parent
Rewrite parentIsObject() function. Currently if a client uploads
a/b/c/d, we always check if c, b, a are actual object or not.

The new code will check with the reverse order and quickly quit if a
segment doesn't exist.

So if a, b, c in 'a/b/c' does not exist in the first place, then returns
false quickly.