listing with prefix prefixed with '/' should be ignored
There are no object keys that can be created on MinIO with
`/' at the beginning of the object name, allowing this is
not required and it behaves as expected with AWS S3.

In AWS S3 if an object is uploaded as `foo/1.txt` cannot be
listed via `prefix=/foo/` although we were allowing it this
is not a correct implementation, only objects that have
`/foo/1.txt` can be listed.

fixes #11265