fix: preserve cache calculated previously while moving from v2 to v3
This ensures that all the prometheus monitoring and usage
trackers to avoid alerts configured, although we cannot
support v1 to v2 here - we can v2 to v3.

In future whenever we make changes such as these, we need
to make sure to preserve older data structure to decode
existing values.