Fix make target hotfix
The code inside the `hotfix` taget is overriding the values set at the
beginning of the Makefile affecting other make targets as well.

For example, running `TAG=mytag make docker` also ends up tagging the
docker image as a hotfix instead of `mytag`.

Using the `eval` function inside the `hotfix` target fixes this.