Add error field instead of using pointers
Remove `omitempty` from the bucket/object count fields and add an
`error` field. When the count cannot be fetched because of any error,
the `error` field will be populated with the actual error message. (Note
that the count will also get populated with the default value of `0`).

The client should read the `count` value only if there is no error.