compression: mark ETag of compressed objects as multipart ETag
This commit ensures that the S3 ETag returned to the client is
always an S3 multipart ETag. An S3 multipart ETag has a `-X`
(`X` being the number of object parts) suffix.

The current behavior of the compression is to mark the
ETag of a compressed object as multipart ETag. This behavior
should be preserved for compressed + encrypted objects.