Fix incorrect marker handling
The S3 specifies:

> Specifies the key to start with when listing objects in a bucket.

MinIO has used the more sensible approach that the marker indicates which is the first key to start listing *after*.

While this is more reasonable it creates subtle incompatibilities.

This is rarely an issue since ListXXX has fields to easily move on to the next objects.

However for manually crafted markers this produces different results to S3.