Avoid synchronizing usage writes
If the periodic `case <-t.C:` save gets held up for a long time it will end up synchronize all disk writes for saving the caches.

We add jitter to per set writes so they don't sync up and don't hold a lock for the write, since it isn't needed anyway.

If an outage prevents writes for a long while we also add individual waits for each disk in case there was a queue.

Furthermore limit the number of buffers kept to 2GB, since this could get huge in large clusters. This will not act as a hard limit, but should be enough for normal operation.