Avoid returning duplicated prefixes containing stale files
When a prefix is present in multiple sets and multiple zones, and it
only contains stale files, it will be shown as empty.

This fix reduces the chance of these use case appearing by testing on
set index & zone index of a found prefix:

If a prefix does not have a quorum in a given set, just do not attribute
it a quorum and push it again to visit it later since there is a chance
that it can have a quorum in another set or zone.

Now if the same prefix does not have quorum in any set in any zone, it
will never have a quourm and won't be shown.

But sometimes a prefix can have quorum because it contains stale
objects, this use case is not fixed.