add SSE-KMS support for single-part operations
This commit adds basic SSE-KMS support for S3
single-part operations (GET,HEAD,PUT) and the
web upload/download functions.

Now, a client can specify the SSE-KMS headers
(algorithm, optional key-id, optional context)
such that the object gets encrypted using the
SSE-KMS method.

This is a minimal implementation that is enabled
only for single part operations. Support for
multipart operations will be added in a separate
PR.

Further, this commit does not try to do any refactoring
and instead tries to implement SSE-KMS as a minimal
change to the code base. However, refactoring the entire
crypto-related code is planned - but needs a separate
effort.