fix: bucket / object count and size returned as 0
The ServerHealthInfo was first populating healthInfoMessage by calling
adm.ServerInfo() and then decoding the JSON response of /healthinfo into
the same struct. In this process, the uint values inside the Buckets,
Objects and Usage structs were getting overwritten with zero.

Reversed the logic by calling adm.ServerInfo() after /healthinfo to
avoid this and ensure that the bucket/object data is retained in the
health report.