Listing: Do not include marker in CommonPrefixes
Walk() and Merge code can return dir__XLDIR__ as last element in a page
list and dir__XLDIR__ as the first element in the next list page.

dir__XLDIR__ is shown in the second page list because the marker set to
dir/ is meant to skip dir/ and not dir__XLDIR__

To fix this, the code will avoid adding the marker itself in the listing
result.