Fix hanging erasure writes
When parallelWriter detects and error it nils out the `p.writers[i]`.

However, this slice is also used for closing the writers, so close is never called on these.

Furthermore when an error is returned from a write it is now reported to the reader.

bonus: remove unused heal param from `newBitrotWriter`.