[Tiering] Support remote tiers with object versioning
- Adds versioning support for S3 based remote tiers that have versioning
enabled. This ensure that when reading or deleting we specify the specific
version ID of the object. In case of deletion this is important to ensure that
the object version is actually deleted instead of simply being marked for
deletion.

- Stores the remote object's version id in the tier-journal. Tier-journal file
version is not bumped up as serializing the new jentry struct version is
compatible with old journals containing jentry's without the remote object
version id.

- `storageRESTVersion` is bumped up as FileInfo struct now includes a
`TransitionRemoteVersionID` member.

- Azure and GCS support for this feature will be added subsequently.