Add support for existing object replication.
Also adding a API to allow resyncing replication when
existing object replication is enabled and the remote target
is entirely lost. With `mc replicate reset` command, the
objects that are eligible for replication as per the replication
config will be resynced to target if existing object replication
is enabled on the rule.