Revert heal locks
A lot of healing is likely to be on non-existing objects and locks are very expensive and will slow down scanning significantly.

In cases where all are valid or all are broken allow rejection without locking.

Keep the existing behaviour, but move the check for dangling objects to after the lock has been acquired.

```
	_, err = getLatestFileInfo(ctx, partsMetadata, errs)
	if err != nil {
		return er.purgeObjectDangling(ctx, bucket, object, versionID, partsMetadata, errs, []error{}, opts)
	}
```

Revert "heal: Hold lock when reading xl.meta from disks (#12362)"

This reverts commit abd32065aaae4080396a1b4b04a110454368b028.