Require clear header naming for prefix deletion
DELETE a bucket & object uses similar HTTP requests. It is possible that
a user playing with lowlevel requests to send a DELETE request with
x-minio-force-delete with an empty object name, which means the whole
bucket could be removed accidentely.

Make the intention more clear in headers by renaming x-minio-force-delete to
x-minio-force-bucket-deletion and x-minio-force-prefix-deletion.