Fix Multipool Precondition Crosstalk
We are racing when doing multi-pool (*erasureServerPools) GetObjectNInfo() calls.

We are sending opts.CheckPrecondFn to all pools. This function writes to the response writer.

That means if one pool fails we write the error and the response code to w, while we move on to another pool.

This can lead to wrong writes in cases in cases where objects exist on multiple pools.

Modify `CheckPrecondFn` to return an error that can later be applied to the writer, or skipped if not interesting.